restarted to see if I could find the error with calculating guesses
